Move internal/maputil to daemon/libnetwork/internal/maputil
